split OS RNG into rand_os
