+++ rand_os/README.md
@@ -1,0 +1,28 @@
+# rand_os
+
+[![Build Status](https://travis-ci.org/rust-random/rand.svg?branch=master)](https://travis-ci.org/rust-random/rand)
+[![Build Status](https://ci.appveyor.com/api/projects/status/github/rust-random/rand?svg=true)](https://ci.appveyor.com/project/rust-random/rand)
+[![Latest version](https://img.shields.io/crates/v/rand_os.svg)](https://crates.io/crates/rand_os)
+[![Documentation](https://docs.rs/rand_os/badge.svg)](https://docs.rs/rand_os)
+[![Minimum rustc version](https://img.shields.io/badge/rustc-1.22+-yellow.svg)](https://github.com/rust-random/rand#rust-version-requirements)
+[![License](https://img.shields.io/crates/l/rand_os.svg)](https://github.com/rust-random/rand/tree/master/rand_os#license)
+
+A random number generator that retrieves randomness straight from the
+operating system.
+
+This crate depends on [rand_core](https://crates.io/crates/rand_core) and is
+part of the [Rand project](https://github.com/rust-random/rand).
+
+Documentation:
+[master branch](https://rust-random.github.io/rand/rand_os/index.html),
+[by release](https://docs.rs/rand_os)
+
+[Changelog](CHANGELOG.md)
+
+## License
+
+`rand_os` is distributed under the terms of both the MIT license and the
+Apache License (Version 2.0).
+
+See [LICENSE-APACHE](LICENSE-APACHE) and [LICENSE-MIT](LICENSE-MIT), and
+[COPYRIGHT](COPYRIGHT) for details.
